requirements/base.txt
celery==4.0.2
kombu==4.1.0
Django==1.11.6
django-braces==1.11.0
django-environ==0.4.3
django-extensions==1.7.9
django-model-utils==3.0.0
Pillow==4.1.1
psycopg2==2.7.1
pytz==2017.2
raven==6.1.0
redis>=2.10.5
django-redis-cache==1.7.1
django-solo==1.1.2
django-silk==1.0.0
django-compressor==2.2
django-oscar==1.5
django-blog-zinnia==0.18.1
django-ckeditor==5.3.0
-e git://github.com/jazzband/django-constance.git#egg=constance
django-picklefield==1.0.0
djangorestframework==3.6.4
django-rest-auth==0.9.2
stripe==2.35.1
cryptography==2.0.3
django-rosetta==0.7.13
django-filter==1.0.4
django-modeltranslation==0.12.1
WeasyPrint==0.41
coreapi==2.3.3
coreschema==0.0.4
pandas==0.23.0
PyPDF2==1.26.0
qrcode==6.1
gevent==1.4.0
pylibdmtx==0.1.9 